72706: Move attachment logic to own file
