+++ modules/ivc_champva/app/models/ivc_champva/vha_10_10d.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
-require 'securerandom'
-
 module IvcChampva
   class VHA1010d
     include Virtus.model(nullify_blank: true)
+    include Attachments
 
     attribute :data
 
     def initialize(data)
       @data = data
       @uuid = SecureRandom.uuid
+      @form_id = 'vha_10_10d'
     end
 
     def metadata
@@ -31,25 +31,6 @@ module IvcChampva
       }
     end
 
-    def handle_attachments(file_path)
-      uuid = @uuid # Generate the UUID as an instance variable
-      file_path_uuid = file_path.gsub('vha_10_10d-tmp', "#{uuid}_vha_10_10d-tmp")
-      File.rename(file_path, file_path_uuid)
-      attachments = get_attachments
-      file_paths = [file_path_uuid]
-
-      if attachments.count.positive?
-        attachments.each_with_index do |attachment, index|
-          new_file_name = "#{uuid}_vha_10_10d-tmp#{index + 1}.pdf"
-          new_file_path = File.join(File.dirname(attachment), new_file_name)
-          File.rename(attachment, new_file_path)
-          file_paths << new_file_path
-        end
-      end
-
-      file_paths
-    end
-
     def submission_date_config
       { should_stamp_date?: false }
     end
@@ -60,23 +41,6 @@ module IvcChampva
 
     def respond_to_missing?(method)
       method == :handle_attachments || super
-    end
-
-    private
-
-    def get_attachments
-      attachments = []
-
-      # TODO: We need to look into generatings individual PDFs for each
-      # attachment based on what PEGA needs
-      supporting_documents = @data['supporting_docs']
-      if supporting_documents
-        confirmation_codes = []
-        supporting_documents&.map { |doc| confirmation_codes << doc['confirmation_code'] }
-        PersistentAttachment.where(guid: confirmation_codes).map { |attachment| attachments << attachment.to_pdf }
-      end
-
-      attachments
     end
   end
 end
