+++ modules/ivc_champva/lib/ivc_champva.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'ivc_champva/engine'
+require 'securerandom'
 
 module IvcChampva
   # Your code goes here...
